fastapi_vue/AIzhujiao/homework-help/ke_knowledge_engineering_tutor/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
openai>=1.52.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-pptx==1.0.2
